emr-etl-project/etl/requirements.txt
PyYAML==6.0
boto3==1.26.0
pyspark==3.5.1
mysql-connector-python==8.1.0